requirements.txt
aiofiles==23.2.1
altair==5.1.0
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.3
fastapi==0.103.0
ffmpy==0.3.1
filelock==3.12.3
fonttools==4.42.1
fsspec==2023.6.0
ftfy==6.1.1
gradio==3.41.2
gradio_client==0.5.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
orjson==3.9.5
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
wcwidth==0.2.6
websockets==11.0.3
